Moved validation to the ControlledSendAmount
