+++ src/components/InputPillGroup/InputWithButtonPillGroup.controlled.tsx
@@ -1,7 +1,4 @@
-import { useCallback } from 'react';
 import { Controller, FieldValues } from 'react-hook-form';
-import { utils } from 'ethers';
-import { fieldsErrors } from '../../constants';
 import { ControlledInputWithButtonPillGroupProps } from './InputWithButtonPillGroup.types';
 import { InputWithButtonPillGroup } from './InputWithButtonPillGroup.component';
 
@@ -14,7 +11,7 @@ export const ControlledInputWithButtonPillGroup = <
   setValue,
   totalAmount,
   totalAmountDecimals,
-  feesAndLimits,
+  validate,
   ...props
 }: ControlledInputWithButtonPillGroupProps<FormValues>) => {
   const onButtonChange = (newValue: string) => {
@@ -22,33 +19,6 @@ export const ControlledInputWithButtonPillGroup = <
       shouldValidate: true,
     });
   };
-
-  const validate = useCallback(
-    (v: string) => {
-      if (!totalAmount) return true;
-      if (utils.parseUnits(v, totalAmountDecimals).gt(totalAmount)) {
-        return fieldsErrors.amountGreaterThanBalance;
-      }
-      if (
-        feesAndLimits?.maxTransfer &&
-        utils.parseUnits(v, totalAmountDecimals).gt(feesAndLimits?.maxTransfer)
-      ) {
-        return fieldsErrors.amountGreaterThanMaxLimit;
-      }
-      if (
-        feesAndLimits?.minTransfer &&
-        utils.parseUnits(v, totalAmountDecimals).lt(feesAndLimits?.minTransfer)
-      ) {
-        return fieldsErrors.amountLessThanMinLimit;
-      }
-    },
-    [
-      feesAndLimits?.maxTransfer,
-      feesAndLimits?.minTransfer,
-      totalAmount,
-      totalAmountDecimals,
-    ]
-  );
 
   return (
     <Controller
