+++ src/pages/Aggregator/SendAmount/SendAmount.component.tsx
@@ -9,10 +9,10 @@ export const SendAmountComponent = <FormValues extends FieldValues>({
   startingTokenBalanceState,
   startingTokenBalance,
   startingTokenDecimals,
-  feesAndLimits,
   name,
   control,
   setValue,
+  validate,
 }: SendAmountComponentProps<FormValues>) => (
   <>
     <Box sx={{ mb: 8, position: 'relative' }}>
@@ -31,9 +31,9 @@ export const SendAmountComponent = <FormValues extends FieldValues>({
       symbol={startingTokenName}
       totalAmount={startingTokenBalance}
       totalAmountDecimals={startingTokenDecimals}
-      feesAndLimits={feesAndLimits}
       control={control}
       setValue={setValue}
+      validate={validate}
     />
   </>
 );
