+++ src/pages/Aggregator/SendAmount/SendAmount.container.tsx
@@ -1,5 +1,7 @@
 import { FieldValues } from 'react-hook-form';
 import { useSelector } from 'react-redux';
+import { useCallback } from 'react';
+import { utils } from 'ethers';
 import { SendAmountComponent } from './SendAmount.component';
 import {
   feesAndLimitsSelector,
@@ -8,6 +10,7 @@ import {
   startingTokenDecimalsSelector,
 } from '../../../store/aggregator/aggregator.selectors';
 import { SendAmountContainerProps } from './SendAmount.types';
+import { fieldsErrors } from '../../../constants';
 
 export const SendAmount = <FormValues extends FieldValues>(
   props: SendAmountContainerProps<FormValues>
@@ -17,14 +20,36 @@ export const SendAmount = <FormValues extends FieldValues>(
   const startingTokenBalanceState = useSelector(
     startingTokenBalanceStateSelector
   );
-  const feesAndLimits = useSelector(feesAndLimitsSelector);
+  const { minTransfer, maxTransfer } = useSelector(feesAndLimitsSelector);
+
+  const validate = useCallback(
+    (v: string) => {
+      if (!startingTokenBalance) return true;
+      if (utils.parseUnits(v, startingTokenDecimals).gt(startingTokenBalance)) {
+        return fieldsErrors.amountGreaterThanBalance;
+      }
+      if (
+        maxTransfer &&
+        utils.parseUnits(v, startingTokenDecimals).gt(maxTransfer)
+      ) {
+        return fieldsErrors.amountGreaterThanMaxLimit;
+      }
+      if (
+        minTransfer &&
+        utils.parseUnits(v, startingTokenDecimals).lt(minTransfer)
+      ) {
+        return fieldsErrors.amountLessThanMinLimit;
+      }
+    },
+    [maxTransfer, minTransfer, startingTokenBalance, startingTokenDecimals]
+  );
 
   return (
     <SendAmountComponent
       startingTokenBalance={startingTokenBalance}
       startingTokenDecimals={startingTokenDecimals}
       startingTokenBalanceState={startingTokenBalanceState}
-      feesAndLimits={feesAndLimits}
+      validate={validate}
       {...props}
     />
   );
