+++ src/pages/Aggregator/SendAmount/SendAmount.types.tsx
@@ -1,7 +1,6 @@
 import { UseFormSetValue } from 'react-hook-form';
 import { ControlledInputType } from '../../../components/TextInput/TextInput.types';
 import { TokenEnum } from '../../../config/tokens';
-import { FeesAndLimitsType } from '../../../store/aggregator/aggregator.state';
 import { FiniteStates } from '../../../utils/types';
 
 export type SendAmountContainerProps<FormValues> = {
@@ -14,7 +13,7 @@ export type TokenBalanceProps = {
   startingTokenBalanceState: FiniteStates;
   startingTokenBalance?: string;
   startingTokenDecimals?: number;
-  feesAndLimits?: FeesAndLimitsType;
+  validate?: (v: string) => string | true | undefined;
 };
 
 export type SendAmountComponentProps<FormValues> =
